Update alma_print_slip.js
In the new Alma layout the JavaScript previously used to get the library name does not work. The simplest fix is to assign your library name to the library variable with a literal string. Thanks to Sue Slivan and her colleague at Finger Lakes Community College for finding the solution.